Refactor: Move download functions to CloudFeed
