+++ handlers/cloudfeed.go
@@ -1,8 +1,10 @@
 package handlers
 
 import (
+	"context"
 	"encoding/json"
 	"net/http"
+	"time"
 
 	"github.com/energietransitie/twomes-backoffice-api/internal/helpers"
 	"github.com/energietransitie/twomes-backoffice-api/services"
@@ -48,6 +50,30 @@ func (h *CloudFeedHandler) Create(w http.ResponseWriter, r *http.Request) error 
 
 		return NewHandlerError(err, "internal server error", http.StatusInternalServerError)
 	}
+
+	return nil
+}
+
+type DownloadArgs struct {
+	AccountID   uint
+	CloudFeedID uint
+	StartPeriod time.Time
+	EndPeriod   time.Time
+}
+
+// Handle RPC endpoint for downloading data from a cloud feed.
+func (h *CloudFeedHandler) Download(args DownloadArgs, reply *string) error {
+	cfa, err := h.service.Find(cloudfeed.CloudFeed{AccountID: args.AccountID, CloudFeedTypeID: args.CloudFeedID})
+	if err != nil {
+		return err
+	}
+
+	err = h.service.Download(context.Background(), cfa, args.StartPeriod, args.EndPeriod)
+	if err != nil {
+		return err
+	}
+
+	*reply = "Downloaded data from cloud feed. Check server logs for more information."
 
 	return nil
 }
