+++ handlers/cloudfeedtype.go
@@ -1,14 +1,11 @@
 package handlers
 
 import (
-	"context"
 	"encoding/json"
 	"net/http"
-	"time"
 
 	"github.com/energietransitie/twomes-backoffice-api/internal/helpers"
 	"github.com/energietransitie/twomes-backoffice-api/services"
-	"github.com/energietransitie/twomes-backoffice-api/twomes/cloudfeed"
 	"github.com/energietransitie/twomes-backoffice-api/twomes/cloudfeedtype"
 	"github.com/sirupsen/logrus"
 )
@@ -45,30 +42,6 @@ func (h *CloudFeedTypeHandler) Create(w http.ResponseWriter, r *http.Request) er
 	if err != nil {
 		return NewHandlerError(err, "internal server error", http.StatusInternalServerError).WithLevel(logrus.ErrorLevel)
 	}
-
-	return nil
-}
-
-type DownloadArgs struct {
-	AccountID   uint
-	CloudFeedID uint
-	StartPeriod time.Time
-	EndPeriod   time.Time
-}
-
-// Handle RPC endpoint for downloading data from a cloud feed.
-func (h *CloudFeedHandler) Download(args DownloadArgs, reply *string) error {
-	cfa, err := h.service.Find(cloudfeed.CloudFeed{AccountID: args.AccountID, CloudFeedTypeID: args.CloudFeedID})
-	if err != nil {
-		return err
-	}
-
-	err = h.service.Download(context.Background(), cfa, args.StartPeriod, args.EndPeriod)
-	if err != nil {
-		return err
-	}
-
-	*reply = "Downloaded data from cloud feed. Check server logs for more information."
 
 	return nil
 }
